local timer, gdt, idt

+++ src/memory/memory.h
@@ -1,14 +1,15 @@
+#include<std.h>
 
 typedef struct MemoryManager MemoryManager;
 
-MemoryManager *createMemoryManager(unsigned base, unsigned size);
+MemoryManager *createMemoryManager(uintptr_t base, size_t size);
 MemoryManager *initKernelMemoryManager(void);
 
-void *allocateAligned(MemoryManager *m, unsigned size, unsigned align);
-void* allocate(MemoryManager* m, unsigned size);
-unsigned getAllocatedSize(MemoryManager* m);
+void *allocateAligned(MemoryManager *m, uintptr_t size, size_t align);
+void* allocate(MemoryManager* m, size_t size);
+size_t getAllocatedSize(MemoryManager* m);
 
 typedef struct PageManager PageManager;
 PageManager *initKernelPageManager(MemoryManager *m);
-unsigned getUsingSize(PageManager *p);
-unsigned getUsableSize(PageManager *p);
+size_t getUsingSize(PageManager *p);
+size_t getUsableSize(PageManager *p);
